Solve the Model <-> Modeler chicken and egg problem with an atom
The problem is that both the models and the modeler need to access
each other, mostly for looking up (other) models, but immutable
data structures make this hard. The solution applied here is to
store the models passed to the modeler in an atom, create the modeler
first, pass it down to model constructors and store the resulting
models in the atom. Now models can access the modeler and can look
up other models through it - e.g. in
brahman.backends.datascript/schema->attrs.