DataScript: Add support for Om Next links in queries
This adds more functionality to the query-store implementation of
the DataScript backend. Prior to converting the Om Next query to
a DataScript query, all links are extracted and stripped from
the Om Next query, remembering the paths they appeared at.

The Om Next query without links is then converted and executed
against DataScript. The links are queried separately one by one.
In the final step, the results of these links queries are merged
back into the DataScript query result according to their
paths  in the original Om Next query.